Fix crash on Exif page due to format error

+++ app/src/main/java/com/siju/acexplorer/main/model/ExifData.kt
@@ -24,7 +24,7 @@ data class ExifData(val context : Context, val tag : String, var value : String)
             }
         }
         else if (tag == ExifInterface.TAG_EXPOSURE_TIME) {
-            value = String.format(Locale.ENGLISH, context.getString(R.string.exposure_sec), value)
+            value = String.format(Locale.ENGLISH, context.getString(R.string.exposure_sec), value.toFloat())
         }
     }
 }
